D3: regenerate program on demand only
We used to regenerate a program when the page was loaded and the user
input was empty. This led to weird behavior, since a page reload was
reusing the previous submission's $_POST, hence a reload after an empty
answer was generating a new program, but a reload after a non-empty
answer was not.

There are now 3 cases where we regenerate a program are now:

- First visit
- Timeout
- Explicit request